Fix for IE warning logic
