+++ apps/site/assets/js/app.js
@@ -38,7 +38,6 @@ import photoGallery from "./photo-gallery";
 import { init as embeddedSearchInit } from "./algolia-embedded-search";
 import * as homepageSearch from "./algolia-homepage-search";
 import * as globalSearch from "./algolia-global-search";
-import AlgoliaResults from "./algolia-results";
 import fullstory from "./fullstory";
 import inputFocus from "./input-focus";
 import channels from "./channels";
@@ -48,7 +47,6 @@ import dismissFullscreenError from "../ts/app/dismiss-fullscreen-error";
 import tripPlannerWidget from "./trip-planner-widget";
 import eventPageSetup from "./event-page-setup";
 import previousEventsButton from "./view-previous-events";
-import IEWarning from "../ts/ie-warning/ie-warning";
 
 document.body.className = document.body.className.replace("no-js", "js");
 
@@ -369,4 +367,3 @@ dismissFullscreenError();
 tripPlannerWidget();
 eventPageSetup();
 previousEventsButton();
-IEWarning();
