+++ apps/site/assets/ts/ie-warning-entry.ts
@@ -1,0 +1,3 @@
+import IEWarning from "./ie-warning/IEWarning";
+
+IEWarning();
