+++ apps/site/assets/webpack.config.base.js
@@ -32,7 +32,8 @@ module.exports = {
     leaflet: ["./ts/leaflet-entry.ts"],
     schedule: ["./ts/schedule-entry.ts"],
     tripplanresults: ["./ts/trip-plan-results-entry.ts"],
-    projects: ["./ts/projects-entry.ts"]
+    projects: ["./ts/projects-entry.ts"],
+    iewarning: ["./ts/ie-warning-entry.ts"]
   },
 
   node: {
